Deleting .github/workflows directory as Action workflow incompatable w/ RPi4 and earlier.
Github Action could not successfully build an image that can be run by an RPi 4 and ealier due to them making use of ARMv8-A which does not support atomic instructions. This support is needed for GH Actions to build image and was added in ARMv8.1-A.

Will use another method to publish images.